% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_seasonality}
\alias{check_seasonality}
\title{Function that returns values from optional seasonal deterministic model}
\usage{
check_seasonality(theta, mpl_pf, season_model)
}
\arguments{
\item{theta}{Posterior distribution of pMCMC parameters. Default = NULL}

\item{mpl_pf}{Model parameter list. Default = NULL}

\item{season_model}{Seasonality model to be used for the optional deterministic
model. Default = NULL}
}
\description{
\code{check_seasonality} Calculates time series of the prefix seasonal deterministic
model given the posterior distribution of pMCMC parameters
}
