% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_u5}
\alias{compare_u5}
\title{Compare function to calculate likelihood}
\usage{
compare_u5(state, observed, pars = NULL)
}
\arguments{
\item{state}{Model output. Default = NULL}

\item{observed}{Oberved data. Default = NULL}

\item{pars}{Parameters, optional.}
}
\description{
\code{compare_u5} Compare function that compares observed data with model estimate
to calculate likelihood for the particle filter. Equates observed prevalence
to prevalence under 5 year olds in the model.
}
