% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_na}
\alias{format_na}
\title{Change missing data to 0s so dbinom function returns 0 (instead of NA)}
\usage{
format_na(df)
}
\arguments{
\item{df}{Data input from user
Default = NULL}
}
\description{
\code{format_na} Change missing data to 0s so dbinom function returns 0 (instead of NA)
}
