% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_init_EIR}
\alias{get_init_EIR}
\title{Return an initial EIR based on a user-given target prevalence in <5 yead old children}
\usage{
get_init_EIR(par, mpl)
}
\arguments{
\item{par}{An initial EIR value from optim function}

\item{mpl}{Model parameter list}
}
\description{
\code{get_init_EIR} Return an initial EIR based on a user-given target prevalence in <5 yead old children
}
