% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_odds_from_prev}
\alias{get_odds_from_prev}
\title{Converts probability (prevalence) to log-odds}
\usage{
get_odds_from_prev(prev)
}
\arguments{
\item{prev}{A probability or prevalence value. Must be between 0 and 1.}
}
\description{
\code{get_odds_from_prev} Converts probability (prevalence) to log-odds
}
