% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prev_from_log_odds}
\alias{get_prev_from_log_odds}
\title{Converts log-odds to probability (or prevalence)}
\usage{
get_prev_from_log_odds(log_odds)
}
\arguments{
\item{log_odds}{A log-odds value.}
}
\description{
\code{get_prev_from_log_odds} Converts log-odds to probability (or prevalence)
}
