% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ll_binom}
\alias{ll_binom}
\title{Binomial function that checks for NAs}
\usage{
ll_binom(positive, tested, model)
}
\arguments{
\item{positive}{Observed number of positive tests. Default = NULL}

\item{tested}{Observed number of tests. Default = NULL}

\item{model}{Output from model.  Default = NULL}
}
\description{
\code{ll_binom} Binomial function that checks for NAs to be used in the compare
function in the particle filter.
}
