% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pmcmc.R
\name{run_pmcmc}
\alias{run_pmcmc}
\title{Run a pMCMC}
\usage{
run_pmcmc(
  data_raw = NULL,
  data_raw_pg = NULL,
  data_raw_mg = NULL,
  init_EIR = 10,
  target_prev = NULL,
  target_prev_group = "5.10",
  n_particles = 200,
  proposal_matrix,
  max_param = 1000,
  prop_treated = 0.4,
  n_steps = 500,
  n_threads = 4,
  n_chains = 1,
  n_workers = 1,
  state_check = FALSE,
  country = NULL,
  admin_unit = NULL,
  seasonality_on = TRUE,
  seasonality_check = FALSE,
  seed = 1L,
  start_pf_time = 30,
  particle_tune = FALSE,
  comparison = c("u5", "pg", "pgmg", "pgsg", "ancall"),
  initial = "informed"
)
}
\arguments{
\item{data_raw}{Time series data to fit model}

\item{init_EIR}{A single value or a dataframe with two columns (time and EIR)
to specify historical malaria transmission levels before
data collection began.}

\item{target_prev}{Return an initial EIR value (from the equilibrium solution),
given a target prevalence in under 5yos}

\item{n_particles}{Number of particles to be used in pMCMC (default = 200)}

\item{proposal_matrix}{Proposal matrix for MCMC parameters}

\item{max_param}{Ceiling for proposed stochastic parameter (either EIR or betaa) values (default = 1000)}

\item{prop_treated}{Proportion of clinical cases that receive effective treatment (default = 40\%)}

\item{n_steps}{Number of MCMC steps in a single chain (default = 500)}

\item{n_threads}{Number of processing threads (default = 4)}

\item{n_chains}{Number of chains (default = 1)}

\item{n_workers}{Number of workers (default = 4)}

\item{state_check}{Run equilibrium checks, if state_check = 1, returns expected deriv values which should equal 0 and sets stochastic model to have EIR constant at init_EIR
If state_check = 1 and seasonality_on = 1, then the deterministic seasonal model is still run, but theta2 is forced to 1, forcing a constant seasonality profile
If state_check = 0, no values are printed}

\item{country}{Name of country (needed if using seasonality model)}

\item{admin_unit}{Name of administrative unit (needed if using seasonality model)}

\item{seasonality_on}{Toggle seasonality model run before observed time period (default = 1)}

\item{seasonality_check}{Toggle saving values of seasonality equilibrium (default = 1)}

\item{seed}{Allows user to specify a seed (default = 1L)}

\item{start_pf_time}{Number of days before first observation that particle filter will start (default = 30)}

\item{particle_tune}{Logical to determine if tuning the number of particles should be performed.}

\item{comparison}{The comparison function to be used. Either 'u5' which
equates the observed prevalence to prevalence under 5 years old in
the model or 'pgmg' which calculates prevalence in primigravid and
multigravid pregnant women for comparison with observed ANC data.}

\item{initial}{Is the initial equilibrium state informed by the user ('user-informed') or by the observed data ('fitted')?}

\item{preyears}{Length of time in years the deterministic seasonal model should run before Jan 1 of the year observations began (default = 2)}
}
\description{
This function sets up and runs a particle MCMC that uses Dust, Odin and MCState
}
