% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pf.R
\name{test_pf}
\alias{test_pf}
\title{Test the particle filter}
\usage{
test_pf(
  data_raw,
  n_particles = 200,
  proposal_matrix,
  max_EIR = 1000,
  max_steps = 1e+07,
  atol = 0.001,
  rtol = 1e-06,
  n_steps = 500,
  n_threads = 4,
  lag_rates = 10,
  state_check = 0,
  country = NULL,
  admin_unit = NULL,
  preyears = 2,
  seasonality_on = 1,
  seasonality_check = 0,
  seed = 1L,
  start_pf_time = 30,
  stoch_param = c("EIR", "betaa")
)
}
\arguments{
\item{data_raw}{Time series data to fit model}

\item{n_particles}{Number of particles to be used in pMCMC (default = 200)}

\item{proposal_matrix}{Proposal matrix for MCMC parameters}

\item{max_EIR}{Ceiling for proposed EIR values (default = 1000)}

\item{max_steps}{Maximum steps for particle filter (default = 1e7)}

\item{atol}{atol for particle filter (default = 1e-3)}

\item{rtol}{rtol for particle filter (default = 1e-6)}

\item{n_steps}{Number of MCMC steps in a single chain (default = 500)}

\item{n_threads}{Number of processing threads (default = 4)}

\item{lag_rates}{Number of delay compartments (default = 10)}

\item{state_check}{Run equilibrium checks, if state_check = 1, returns expected deriv values which should equal 0 and sets stochastic model to have EIR constant at init_EIR
If state_check = 1 and seasonality_on = 1, then the deterministic seasonal model is still run, but theta2 is forced to 1, forcing a constant seasonality profile
If state_check = 0, no values are printed}

\item{country}{Name of country (needed if using seasonality model)}

\item{admin_unit}{Name of administrative unit (needed if using seasonality model)}

\item{preyears}{Length of time in years the deterministic seasonal model should run before Jan 1 of the year observations began (default = 2)}

\item{seasonality_on}{Toggle seasonality model run before observed time period (default = 1)}

\item{seasonality_check}{Toggle saving values of seasonality equilibrium (default = 1)}

\item{seed}{Allows user to specify a seed (default = 1L)}

\item{start_pf_time}{Number of days before first observation that particle filter will start (default = 30)}
}
\description{
This function sets up and runs a particle MCMC that uses Dust, Odin and MCState
}
