% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform}
\alias{transform}
\title{Transformation function that calculates initial values for stochastic model}
\usage{
transform(mpl_pf, season_model)
}
\arguments{
\item{season_model}{Seasonality model to be used for the optional deterministic
model. Default = NULL}

\item{mpl}{Model parameter list. Default = NULL}
}
\description{
\code{transform} Calculates the model equilibrium based on an initial EIR values,
then optionally runs a deterministic seasonal model and returns initial
values to be used for the stochastic model fitting.
}
