% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_init}
\alias{transform_init}
\title{transform final state of a seasonal model into the initial state of a stochastic model}
\usage{
transform_init(final_state = NULL)
}
\arguments{
\item{final_state}{Final time point of a seasonal model
Default = NULL}
}
\description{
\code{transform_init} Transform final state of a seasonal model into the initial state of a stochastic model
}
