% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_process}
\alias{data_process}
\title{transform final state of a seasonal model into the initial state of a stochastic model}
\usage{
data_process(data_raw = NULL, start_pf_time = NULL)
}
\arguments{
\item{data_raw}{Data input from user
Default = NULL}

\item{start_pf_time}{Amount of time before first observation to start particle filter}
}
\description{
\code{data_process} Transform data file
}
