% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialise}
\alias{initialise}
\title{Estimate the initial state given user inputs}
\usage{
initialise(init_EIR, mpl, det_model)
}
\arguments{
\item{mpl}{Model parameter list. Default = NULL}

\item{season_model}{Seasonality model to be used for the optional deterministic
model. Default = NULL}
}
\description{
\code{transform} Calculates the model equilibrium based on an initial EIR values,
then optionally runs a deterministic seasonal model and returns initial
values to be used for the stochastic model fitting.
}
